+++ testfiles/calls.py
@@ -1,0 +1,4 @@
+print('Hello world')
+
+foo()
+foo()(bar)